готово

+++ test2/Reviews.cs
@@ -25,7 +25,7 @@ namespace test2
         public int UserId { get; set; }
         public Nullable<System.DateTime> ReviewDate { get; set; }
         public string Comment { get; set; }
-        public Nullable<int> OverallRating { get; set; }
+        public Nullable<double> OverallRating { get; set; }
     
         public virtual Hotels Hotels { get; set; }
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
